lc-u d-144 bipartiate graph eg- better mem

+++ leetcode/possible_bipartition.cpp
@@ -3,17 +3,17 @@ public:
     bool possibleBipartition(int n, vector<vector<int>>& dislikes) {
         // example of bipartiate graph
         // color scheme -> 1(red) & -1(black)
-        vector<int> color(n+1,0);
+        vector<short> color(n+1,0);
         vector<bool> visited(n+1,false);
 
-        vector<vector<int>> adj(n+1);
+        vector<vector<short>> adj(n+1);
 
         for(auto& e : dislikes){
             adj[e[0]].push_back(e[1]);
             adj[e[1]].push_back(e[0]);
         }
-        queue<int> q;
-        for(int i=1;i<n+1;i++){
+        queue<short> q;
+        for(short i=1;i<n+1;i++){
             if(!visited[i]){
                 color[i]=1;
                 q.push(i);
@@ -23,7 +23,7 @@ public:
                     if(visited[x]) continue;
 
                     visited[x] = true;
-                    for(int nei : adj[x]){
+                    for(short nei : adj[x]){
                         if(color[nei] == color[x]) return false;
                         color[nei] = -color[x];
                         q.push(nei);
